Move app into dir
