+++ todo_app/static/img/remove.png
