+++ todo_app/static/img/tick.png
