+++ todo_app/tests/ViewModel_test.py
@@ -1,7 +1,7 @@
 import pytest
-from ViewModel import ViewModel
-from Item import Item
-from Status import Status
+from todo_app.ViewModel import ViewModel
+from todo_app.Item import Item
+from todo_app.Status import Status
 import datetime
 
 items = [Item('1', 'do the thing', 'To Do', '', '', ''),
